SynQ compiler: widen Literal::Number from u64 to u128
Literals larger than u64::MAX (e.g. u128::MAX = 2^128-1) were
silently parsed as 0 via unwrap_or(0), causing the overflow test
contract to return I32(1) instead of a RuntimeError.

- ast.rs: Number(u64) -> Number(u128)
- parser.rs: parse() -> parse::<u128>() for number_literal rule
- codegen.rs: threshold comparison updated to i32::MAX as u128

Now running overflow_test.synq produces:
  UInt256 overflow on Add

+++ SynQ/compiler/src/ast.rs
@@ -114,6 +114,6 @@ pub enum Type {
 #[derive(Debug, PartialEq, Clone)]
 pub enum Literal {
     String(String),
-    Number(u64),
+    Number(u128),  // u128 so literals up to UInt256 range parse correctly
     Bool(bool),
 }
